Add skills deletion

+++ app/src/main/java/edu/kmaooad/capstone23/competences/handlers/DeleteSkillHandler.java
@@ -1,6 +1,7 @@
 package edu.kmaooad.capstone23.competences.handlers;
 
 import edu.kmaooad.capstone23.common.CommandHandler;
+import edu.kmaooad.capstone23.common.ErrorCode;
 import edu.kmaooad.capstone23.common.Result;
 import edu.kmaooad.capstone23.competences.commands.DeleteSkill;
 import edu.kmaooad.capstone23.competences.dal.Skill;
@@ -8,6 +9,7 @@ import edu.kmaooad.capstone23.competences.dal.SkillsRepository;
 import edu.kmaooad.capstone23.competences.events.SkillDeleted;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.inject.Inject;
+import org.bson.types.ObjectId;
 
 @RequestScoped
 public class DeleteSkillHandler implements CommandHandler<DeleteSkill, SkillDeleted> {
@@ -19,10 +21,18 @@ public class DeleteSkillHandler implements CommandHandler<DeleteSkill, SkillDele
 
     @Override
     public Result<SkillDeleted> handle(DeleteSkill command) {
-        Skill skill = repository.findById(command.getId());
+        ObjectId id = command.getId();
+        Skill skill = repository.findById(id);
 
-        System.out.println(skill);
+        if (skill == null) {
+            return new Result<>(ErrorCode.NOT_FOUND, "Skill not found");
+        }
 
+        // let's see if it has any children. We won't delete a skill that has children
+        var allChildren = repository.findChildRepositories(id);
+        if (allChildren.size() > 0) {
+            return new Result<>(ErrorCode.CHILDREN_PRESENT, "Skill has children");
+        }
 
         repository.delete(skill);
